Symptoms initial commit
