+++ composeApp/src/commonMain/kotlin/com/healthanalytics/android/presentation/screens/health/SymptomsList.kt
@@ -1,0 +1,14 @@
+package com.healthanalytics.android.presentation.screens.health
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+
+@Composable
+fun SymptomsList() {
+    Column(modifier = Modifier.fillMaxSize()) {
+
+    }
+}
